requirements.txt
langchain-openai==0.0.5
langchain-chroma==0.0.3
langchain-core==0.1.19
langgraph==0.0.24
langchain==0.1.1
gradio==4.16.0
python-dotenv==1.0.0
chromadb==0.4.22
openai==1.12.0
pydantic==2.6.1
tiktoken==0.5.2